restapi

+++ DemoRestAPI/src/main/java/com/example/demorestapi/Controller/ProductController.java
@@ -1,12 +1,10 @@
 package com.example.demorestapi.Controller;
 
-import com.example.demorestapi.Model.Category;
 import com.example.demorestapi.Model.Product;
 import com.example.demorestapi.Service.CategoryService;
 import com.example.demorestapi.Service.ProductService;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -27,53 +25,43 @@ public class ProductController {
     }
 
     @GetMapping("/products/{id}")
-    public Product getProductDetails(@PathVariable("id") Long id) {
-        return productService.getProductById(id);
+    public ResponseEntity<?> getProductDetails(@PathVariable Long id) {
+        Product product = productService.getProductById(id);
+        if (product == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Sản phẩm không tồn tại.");
+        }
+        return ResponseEntity.ok(product);
     }
 
-    @PostMapping("/products/save")
-    public Product saveProduct(@RequestBody Product product) {
-        return productService.saveProduct(product);
+    @PostMapping("/products/add")
+    public ResponseEntity<?> createProduct(@RequestBody Product product) {
+        boolean isCodeExists = productService.isProductCodeExists(product.getProductCode());
+        if (isCodeExists) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Mã sản phẩm đã tồn tại.");
+        }
+        Product savedProduct = productService.saveProduct(product);
+        return ResponseEntity.ok(savedProduct);
     }
+
+    @PutMapping("/products/edit/{id}")
+    public ResponseEntity<?> updateProduct(@PathVariable Long id, @RequestBody Product product) {
+        boolean isProductExists = productService.existsProductById(id);
+        if (!isProductExists) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Sản phẩm không tồn tại.");
+        }
+        product.setId(id);
+        Product savedProduct = productService.saveProduct(product);
+        return ResponseEntity.ok(savedProduct);
+    }
+
 
     @DeleteMapping("/products/delete/{id}")
-    public String deleteProduct(@PathVariable Long id) {
-        productService.deleteProductById(id);
-        return "thanh cong";
-    }
-
-    /*@GetMapping("/products/add")
-    public String showAddProductForm(Model model) {
-        List<Category> categories = productService.getAllCategories();
-        model.addAttribute("categories", categories);
-        model.addAttribute("product", new Product());
-        return "add-product";
-    }
-
-    @PostMapping("/products")
-    public String addProduct(@ModelAttribute("product") Product product) {
-        productService.saveProduct(product);
-        return "redirect:/products";
-    }
-
-    @GetMapping("/products/edit/{id}")
-    public String showEditProductForm(@PathVariable("id") Long id, Model model) {
-        Product product = productService.getProductById(id);
-        if (product != null) {
-            List<Category> categories = productService.getAllCategories();
-            model.addAttribute("categories", categories);
-            model.addAttribute("product", product);
-            return "product-edit";
-        } else {
-            return "redirect:/products";
+    public ResponseEntity<String> deleteProduct(@PathVariable Long id) {
+        boolean exists = productService.existsProductById(id);
+        if (!exists) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("sản phầm không tồn tại");
         }
+        productService.deleteProductById(id);
+        return ResponseEntity.ok("Xóa thành công.");
     }
-
-    @PostMapping("/products/{id}")
-    public String updateProduct(@PathVariable("id") Long id, @ModelAttribute("product") Product product) {
-        product.setId(id);
-        productService.saveProduct(product);
-        return "redirect:/products";
-    }*/
 }
-
